backend/requirements.txt
# Core framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
Flask-Login>=0.6.3
Flask-Cors>=4.0.0

# Database & migrations
SQLAlchemy>=2.0.21
alembic>=1.13.1
psycopg2-binary>=2.9


# Password hashing / security
Werkzeug>=3.0.1

# Utilities
python-dotenv>=1.0.1
six>=1.16.0
typing-extensions>=4.10.0
